+++ src/python/scripts/uic.py
@@ -1,8 +1,10 @@
 import sys
 
-from utils import is_installed, remove_vendor_suffix
+from common import (QtToolModuleNotFoundError, catch_known_errors,
+                    is_installed, remove_vendor_suffix)
 
-if __name__ == '__main__':
+
+def run():
     if is_installed('PySide6'):
         from PySide6.scripts.pyside_tool import uic as main
     elif is_installed('PyQt6'):
@@ -10,8 +12,10 @@ if __name__ == '__main__':
     elif is_installed('PyQt5'):
         from PyQt5.uic.pyuic import main
     else:
-        raise ModuleNotFoundError(
-            'No Qt for Python module with uic tool is installed.'
-        )
+        raise QtToolModuleNotFoundError('uic')
     sys.argv[0] = remove_vendor_suffix(sys.argv[0])
     sys.exit(main())
+
+
+if __name__ == '__main__':
+    catch_known_errors(run)